Add files via upload
Monkeypox is an illness caused by the monkeypox virus. It is a viral zoonotic infection, meaning that it can spread from animals to humans. It can also spread from person to person.
It is normally encountered in Central and Western Africa, but has now spread to more than a dozen countries, including the United States